+++ pkg/mod/github.com/labstack/gommon@v0.4.0/go.mod
@@ -1,0 +1,13 @@
+module github.com/labstack/gommon
+
+go 1.12
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/mattn/go-colorable v0.1.11
+	github.com/mattn/go-isatty v0.0.14
+	github.com/stretchr/testify v1.7.0
+	github.com/valyala/fasttemplate v1.2.1
+	golang.org/x/sys v0.0.0-20211103235746-7861aae1554b // indirect
+	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
+)
