+++ pkg/mod/github.com/mattn/go-colorable@v0.1.11/go.mod
@@ -1,0 +1,8 @@
+module github.com/mattn/go-colorable
+
+require (
+	github.com/mattn/go-isatty v0.0.14
+	golang.org/x/sys v0.0.0-20210927094055-39ccf1dd6fa6 // indirect
+)
+
+go 1.13
