+++ pkg/mod/github.com/mattn/go-isatty@v0.0.14/go.mod
@@ -1,0 +1,5 @@
+module github.com/mattn/go-isatty
+
+go 1.12
+
+require golang.org/x/sys v0.0.0-20210630005230-0f9fa26af87c
