+++ pkg/mod/github.com/stretchr/testify@v1.8.1/go.mod
@@ -1,0 +1,10 @@
+module github.com/stretchr/testify
+
+go 1.13
+
+require (
+	github.com/davecgh/go-spew v1.1.1
+	github.com/pmezard/go-difflib v1.0.0
+	github.com/stretchr/objx v0.5.0
+	gopkg.in/yaml.v3 v3.0.1
+)
