+++ pkg/mod/github.com/valyala/fasttemplate@v1.2.1/go.mod
@@ -1,0 +1,5 @@
+module github.com/valyala/fasttemplate
+
+go 1.12
+
+require github.com/valyala/bytebufferpool v1.0.0
