+++ pkg/mod/golang.org/x/crypto@v0.5.0/go.mod
@@ -1,0 +1,11 @@
+module golang.org/x/crypto
+
+go 1.17
+
+require (
+	golang.org/x/net v0.5.0
+	golang.org/x/sys v0.4.0
+	golang.org/x/term v0.4.0
+)
+
+require golang.org/x/text v0.6.0 // indirect
