+++ pkg/mod/golang.org/x/net@v0.5.0/go.mod
@@ -1,0 +1,9 @@
+module golang.org/x/net
+
+go 1.17
+
+require (
+	golang.org/x/sys v0.4.0
+	golang.org/x/term v0.4.0
+	golang.org/x/text v0.6.0
+)
