+++ pkg/mod/golang.org/x/sys@v0.4.0/go.mod
@@ -1,0 +1,3 @@
+module golang.org/x/sys
+
+go 1.17
