+++ pkg/mod/golang.org/x/text@v0.6.0/go.mod
@@ -1,0 +1,10 @@
+module golang.org/x/text
+
+require golang.org/x/tools v0.1.12 // tagx:ignore
+
+require (
+	golang.org/x/mod v0.6.0-dev.0.20220419223038-86c51ed26bb4 // indirect tagx:ignore
+	golang.org/x/sys v0.0.0-20220722155257-8c9f86f7a55f // indirect tagx:ignore
+)
+
+go 1.17
