+++ pkg/mod/gopkg.in/yaml.v3@v3.0.1/go.mod
@@ -1,0 +1,5 @@
+module "gopkg.in/yaml.v3"
+
+require (
+	"gopkg.in/check.v1" v0.0.0-20161208181325-20d25e280405
+)
