+++ pkg/sumdb/sum.golang.org/latest
@@ -1,5 +1,5 @@
 go.sum database tree
-14852015
-Qepxwj2zOZOqZIO9AWmJPzY/sgFFPPfK9jb1fPPOoyk=
+14906819
+J4nqL3N4pLBg9LtF6dzTX8rV6wMnIDA4/Ac1xDQ54tE=
 
-— sum.golang.org Az3grjpvawSHSXh5FTW8TE9HJ4FUaitZlW4tawmKXP9TYI/Rrj8+zANSxpY8Vu5ZPdohfK+4jOmPu8zTojnS27Cixgo=
+— sum.golang.org Az3grqA2iL2Liv7J3lrs5cf5SoPO6aPvXJf5ZMgPwmkXCs7boyWYEsBovor5qOPwoMZXB1hXYzOdFpNK8jtZZSF60As=
