+++ scrapper/home.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="kr">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, inital-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Go Jobs</title>
+</head>
+<body>
+    <h1>Go Job List</h1>
+    <h3>Saramin scrapper</h3>
+    <form method="post" action="/scrape">
+        <input placeholder="What list do u want" name="term" />
+        <button>Search</button>
+    </form>
+</body>
+</html>
